Create bot.html
This chatbot doesn't require any login or authentication, and it uses the ChatterBot library to generate responses based on the user's input. Users can type messages into the input field and click the "Send" button to see the bot's response.
Note that this chatbot is a simple example and doesn't have the same level of sophistication as a chatbot built with a natural language processing API like Dialogflow or Watson.